image-processor/requirements.txt
google-cloud-storage==2.18.2
pillow==11.1.0